Segmented methods and extensions refactored.
